Add Markdown Link Checking (#7)
# Pull Request

## Description

This pull request introduces a new workflow to check for broken Markdown
links and adds the necessary configuration for the link checker. The
most important changes include the addition of a new job in the GitHub
Actions workflow and the configuration file for the link checker.

New workflow additions:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R34-R51):
Added a new job `check-markdown-links` to the workflow, which uses the
`UmbrellaDocs/action-linkspector` action to check for broken Markdown
links.

Configuration for link checker:

*
[`.github/other-configurations/.linkspector.yml`](diffhunk://#diff-f9691f23ea6c4c34ecbaa23a98721e0aa3c7a94f2fcdd38f7b99fe863c5f6305R1-R5):
Added configuration file specifying directories to check and the status
codes considered as alive.