Dependabot for Terraform Modules (#12)
# Pull Request

## Description

This pull request includes updates to the `.github/dependabot.yml` file
to enhance the automation of dependency updates. The changes include the
removal of a commit message prefix for GitHub Actions updates and the
addition of a new configuration for Terraform updates.

Changes to `.github/dependabot.yml`:

* Removed the commit message prefix for GitHub Actions updates to
simplify commit messages.
* Added a new configuration for Terraform updates, including scheduling
details and grouping patterns for updates.