Add depup Repository (#16)
# Pull Request

## Description

This pull request includes changes to the `TODO.md` file and the
addition of a new Terraform configuration for a GitHub repository. The
most important changes are listed below:

Updates to `TODO.md`:

* Removed the `depup` item from the list of tasks.

Addition of Terraform configuration:

* Added a new Terraform resource definition for the `depup` GitHub
repository in `depup.tf`, including settings for repository visibility,
features, and pull request settings.