Add development-ideas Repository  (#17)
# Pull Request

## Description

This pull request includes updates to the `TODO.md` file and the
addition of a new Terraform configuration file for the
`development-ideas` repository. The most important changes include the
removal of the `development-ideas` repository from the `TODO.md` list,
the addition of new tasks to the `TODO.md` file, and the creation of a
new Terraform resource for the `development-ideas` repository.

Changes to `TODO.md`:

* Removed `development-ideas` from the list of repositories to add.
* Added new tasks under a new "Other" section, including adding a
`README.md` and testing the Terraform configuration.

Addition of Terraform configuration:

* Created a new Terraform resource `github_repository` for the
`development-ideas` repository with specific settings for pull requests,
GitHub Pages, and other repository settings.