Add tech-radar Repository (#19)
# Pull Request

## Description

This pull request includes changes to the `TODO.md` and `tech-radar.tf`
files to update project statuses and configure a new GitHub repository
for the Tech Radar project. The most important changes include the
removal of a project from the TODO list and the addition of a
comprehensive configuration for the new GitHub repository.

Updates to project statuses:

*
[`TODO.md`](diffhunk://#diff-5c6a1301c6b59b30a040d747d065e861d3dd98bde0e5a4356d92d594e9835986L17):
Removed the `tech-radar` project from the list, indicating its
completion or removal from the current to-do items.

GitHub repository configuration:

*
[`tech-radar.tf`](diffhunk://#diff-526a89b98c3c3a561938742a9955b448f9eda2305f3e8e78ed31f691aaf1b7caR1-R35):
Added a new resource configuration for the `tech-radar` GitHub
repository, including settings for repository features, pull request
settings, and GitHub Pages configuration.