Add development-environment Repository (#21)
# Pull Request

## Description

This pull request includes changes to the `TODO.md` file and the
addition of a new Terraform resource for the development environment
repository. The most important changes are:

Updates to `TODO.md`:

* Removed the `development-environment` entry from the list.

Addition of Terraform resource:

* Added a new `github_repository` resource for `development-environment`
in `development-environment.tf` with specific settings for repository
visibility, pull request settings, and other repository configurations.