Add python-practice Repository (#24)
# Pull Request

## Description

This pull request makes significant changes to the repository
configuration and project documentation. The most important changes
include the removal of the `python-practice` project from the `TODO.md`
file and the addition of a new Terraform configuration for the
`python-practice` GitHub repository.

Changes to project documentation:

*
[`TODO.md`](diffhunk://#diff-5c6a1301c6b59b30a040d747d065e861d3dd98bde0e5a4356d92d594e9835986L10):
Removed the `python-practice` project from the list.

Repository configuration updates:

*
[`python-practice.tf`](diffhunk://#diff-51b86d1e32e35d1c5637af8d100261389693c2b25b0aba22161f9320dc9c3f5dR1-R24):
Added a new Terraform configuration to create and manage the
`python-practice` GitHub repository with specific settings such as
visibility, repository features, pull request settings, and other
configurations.