Add project-status-checker Repository (#26)
# Pull Request

## Description

This pull request includes updates to the `TODO.md` file and the
addition of a new Terraform resource for the `project-status-checker`
repository. The most important changes are as follows:

Updates to documentation:

*
[`TODO.md`](diffhunk://#diff-5c6a1301c6b59b30a040d747d065e861d3dd98bde0e5a4356d92d594e9835986L10):
Removed the `project-status-checker` entry from the list of tasks.

Infrastructure as Code:

*
[`project-status-checker.tf`](diffhunk://#diff-dd8da8ad925e7c65eaad33d1085e8bd3d4f9f2db8d53d6870a7d2de898e20d41R1-R24):
Added a new Terraform resource to manage the `project-status-checker`
GitHub repository with specified settings, including repository features
and pull request settings.