Add RepoSync Repository (#27)
# Pull Request

## Description

This pull request involves the creation of a new GitHub repository
resource and updates to the `TODO.md` file. The most important changes
include the addition of a new repository configuration and the removal
of the `RepoSync` task from the TODO list.

Repository configuration:

*
[`RepoSync.tf`](diffhunk://#diff-e34cc084b02af44e87d74c6237098804af0fab88a92babad856ac4eecc52d8a4R1-R20):
Added a new GitHub repository resource named `RepoSync` with specific
settings for pull requests and other repository attributes.

Documentation update:

*
[`TODO.md`](diffhunk://#diff-5c6a1301c6b59b30a040d747d065e861d3dd98bde0e5a4356d92d594e9835986L9):
Removed the `RepoSync` task from the list, indicating it is no longer a
pending task.