Add create-repository-tasks Repository (#28)
# Pull Request

## Description

This pull request includes changes to the `TODO.md` file and the
addition of a new Terraform configuration file to create a GitHub
repository with specific settings. The most important changes are
summarized below:

### Updates to `TODO.md`:

* Removed the `create-repository-tasks` item from the list.

### Addition of `create-repository-tasks.tf`:

* Added a new Terraform configuration file to create a GitHub repository
named `create-repository-tasks` with specific settings, including
enabling auto-merge, disabling merge and rebase commits, and setting the
repository visibility to public.