Add aws-timing-scripts Repository (#32)
# Pull Request

## Description

This pull request includes updates to the `TODO.md` file and the
addition of a new GitHub repository resource in the
`aws-timing-scripts.tf` file. The most important changes are as follows:

Updates to documentation:

*
[`TODO.md`](diffhunk://#diff-5c6a1301c6b59b30a040d747d065e861d3dd98bde0e5a4356d92d594e9835986L3-L7):
Removed the section for adding repositories, specifically the entries
for `test-project-status-checker` and `aws-timing-scripts`.

Infrastructure as Code changes:

*
[`aws-timing-scripts.tf`](diffhunk://#diff-5df2671438b732b17488a7071a6a4deb349e89d4e86b821cb6c211019bba5289R1-R19):
Added a new GitHub repository resource for `aws-timing-scripts` with
specific settings for pull requests and other repository configurations.