Fix and Improve GitHub Stats Checks (#45)
# Pull Request

## Description

This pull request includes changes to the `stack/github-stats.tf` file
to update the default branch protection settings for the GitHub
repository. The most important changes include the removal of certain
required status checks and the addition of a new code scanning tool.

Updates to default branch protection settings:

* Removed the "Test GitHub Summary" and "Validate Schema" status checks
from the list of required status checks.
* Added "SonarCloud" to the list of required code scanning tools.