Add Repository Descriptions (#61)
# Pull Request

## Description

This pull request includes changes to the descriptions of several GitHub
repositories. These changes provide more informative descriptions for
each repository.

Updated repository descriptions:

*
[`stack/GitHubPulse.tf`](diffhunk://#diff-f5705c387892a1d3f3fb63d26c09f12b915bb70f1a541cd9a42b82e404533371L5-R5):
Updated the description to "A RSS feed reader for GitHub Blog".
*
[`stack/RepoOrchestratorState.tf`](diffhunk://#diff-547f611809e14b7aeaba133efa6ac141fae7fa97db4380b01c751272014b992bL4-R4):
Updated the description to "A state repository for the RepoOrchestrator
Infrastructure as Code".
*
[`stack/aws-timing-scripts.tf`](diffhunk://#diff-34a66fb6f671fea366fe0b87e09ba5e16fe0b5fae92321030aebab0a3f6b38ceL5-R5):
Updated the description to "Python scripts for timing AWS services".
*
[`stack/github-stats.tf`](diffhunk://#diff-4c83331edafb5463d9be52eca7c9e213c0868cddd414c2b7d97d4e1559a70ab1L5-R5):
Updated the description to "A Website to display my Repository
Statistics".
*
[`stack/pls.tf`](diffhunk://#diff-0e4fa6f82344e6b675ff6ac7414659b4868df4d89e84bf2a16bf0e36ad4ccd8dL5-R5):
Updated the description to "pls (Project Listing System) is a cli tool
to change projects".
*
[`stack/source_scan.tf`](diffhunk://#diff-4b3bf164059b483435d7002f084e05adc33a2a964754f30a5ebc79fa8489ee26L5-R5):
Updated the description to "A tool to count technologies and frameworks
used in my repositories".
*
[`stack/status-sentinel.tf`](diffhunk://#diff-0a754bed0bccffa46d2be5cda18d1855448c8dc58b211c3dfc8ebec894f2569fL5-R5):
Updated the description to "A tool to monitor the status of project
websites".
*
[`stack/useful-commands.tf`](diffhunk://#diff-66628c93f78b671ea7f168606ef7642b16e24c99615b28680ec3d9f01a9106b0L5-R5):
Updated the description to "A collection of my useful commands".