Update repo-overseer status checks (#84)
# Pull Request

## Description

This pull request includes changes to the `stack/repo-overseer.tf` file
to update the default branch protection rules. The most important
changes include adding a new CodeQL analysis action for `actions`,
replacing the JavaScript CodeQL analysis with TypeScript, and making
adjustments to the list of required checks.

Updates to default branch protection rules:

* Added "CodeQL Analysis (actions)" to the list of required checks.
* Replaced "CodeQL Analysis (javascript)" with "CodeQL Analysis
(typescript)" in the list of required checks.