Update tech-radar status checks (#85)
# Pull Request

## Description

This pull request includes updates to the `tech-radar.tf` file to
enhance code quality and security checks. The most important changes are
the addition of CodeQL analysis to the list of checks and required code
scanning tools.

Enhancements to code quality and security checks:

*
[`stack/tech-radar.tf`](diffhunk://#diff-492e56ddb82f7c2f0feda77f5a318465a9a443db5775cad33ae6161905eecc50R55-R61):
Added "CodeQL Analysis" to the list of checks.
*
[`stack/tech-radar.tf`](diffhunk://#diff-492e56ddb82f7c2f0feda77f5a318465a9a443db5775cad33ae6161905eecc50R55-R61):
Included "CodeQL" in the `required_code_scanning_tools` array.