Add useful-commands Default Branch Protection (#90)
# Pull Request

## Description

This pull request includes updates to the Terraform configuration files
to enhance branch protection and status checks for GitHub repositories.
The most important changes are the addition of a new module for default
branch protection and the removal of a duplicated status check entry.

Enhancements to branch protection:

*
[`stack/useful-commands.tf`](diffhunk://#diff-66628c93f78b671ea7f168606ef7642b16e24c99615b28680ec3d9f01a9106b0R26-R43):
Added a new module `useful-commands_default_branch_protection` which
includes required status checks and code scanning tools for the
`useful-commands` repository.

Code quality improvements:

*
[`stack/project-status-checker.tf`](diffhunk://#diff-2a4b5a746b5b79e9a3d33092afd952c2b24b3362b4204558c86fd6b917efe8dfL48):
Removed a duplicated entry for "Check Pull Request Title" in the list of
required status checks.