Update status-sentienl status checks (#92)
This pull request includes changes to the `stack/status-sentinel.tf`
file to update the default branch protection settings. The most
important change is the modification of the CodeQL analysis checks.

Updates to default branch protection settings:

*
[`stack/status-sentinel.tf`](diffhunk://#diff-0a754bed0bccffa46d2be5cda18d1855448c8dc58b211c3dfc8ebec894f2569fL57-R58):
Changed CodeQL analysis from "javascript" to "actions" and added
"typescript" to the list of checks.