Update GitHubPulse Status Checks (#99)
# Pull Request

## Description

This pull request includes an update to the
`GitHubPulse_default_branch_protection` module in the
`stack/GitHubPulse.tf` file to enhance code quality and security checks.

Enhancements to code quality and security checks:

*
[`stack/GitHubPulse.tf`](diffhunk://#diff-f5705c387892a1d3f3fb63d26c09f12b915bb70f1a541cd9a42b82e404533371R45-R49):
Added "CodeQL Analysis" to the list of required status checks and
included "CodeQL" in the `required_code_scanning_tools`.