Update python-practice Status Checks (#100)
# Pull Request

## Description

This pull request includes a change to the `stack/python-practice.tf`
file to improve the clarity of the `CodeQL Analysis` checks by
specifying the languages they apply to.

*
[`stack/python-practice.tf`](diffhunk://#diff-1a54e41d03518683485b02101e0364624e00cd74300a32a95093acc7f483306dL46-R47):
Updated the `module "python-practice_default_branch_protection"` to
specify `CodeQL Analysis (actions)` and `CodeQL Analysis (python)`
instead of the general `CodeQL Analysis`.