Update screenshot_mailinator_email Status Checks (#101)
# Pull Request

## Description

This pull request includes a small change to the
`stack/screenshot_mailinator_email.tf` file. The change updates the list
of actions to include separate entries for CodeQL Analysis for both
`actions` and `python`.

*
[`stack/screenshot_mailinator_email.tf`](diffhunk://#diff-9a9d85335e669478ae9ba161a18f43bd7ae0766645d8506de4ef0aed75e6197eL45-R46):
Updated the list of actions to include "CodeQL Analysis (actions)" and
"CodeQL Analysis (python)" instead of the single "CodeQL Analysis"
entry.