Upgrade GitHub Provider (#104)
# Pull Request

## Description

This pull request includes updates to the GitHub provider version in the
Terraform configuration files. The most important changes are:

Provider version updates:
*
[`modules/default-branch-protection/version.tf`](diffhunk://#diff-083dcdd2848fb453f0966cc360d8160b73b9759560863e07dc0e7ae21f851b8aL6-R6):
Updated the `github` provider version from `6.4.0` to `6.5.0`.
*
[`stack/.terraform.lock.hcl`](diffhunk://#diff-a977c47fa87684d79003fe5b838ec4f35f5d4437ef888396ae59474faa56cfc0L5-R22):
Updated the `github` provider version and constraints from `6.4.0` to
`6.5.0`, and updated the associated hashes.
*
[`stack/terraform.tf`](diffhunk://#diff-071e1b4d3f4377396546dc5bc41672a414c0723beedc4b4ed2a37a2c5086ccf5L6-R6):
Updated the `github` provider version from `6.4.0` to `6.5.0`.