Update repo-overseer status checks (#112)
# Pull Request

## Description

This pull request includes changes to the
`repo-overseer_default_branch_protection` module in the
`stack/repo-overseer.tf` file. The most important changes are the
reordering and addition of code checks.

Changes to code checks:

*
[`stack/repo-overseer.tf`](diffhunk://#diff-83b44e78de7c37d798c938495c4a000f8a5cb8a25c8bb4f687ca44b583a68d4aL57-R64):
Removed "Run TypeScript Code Checks" from its original position and
added it back in a new position.
*
[`stack/repo-overseer.tf`](diffhunk://#diff-83b44e78de7c37d798c938495c4a000f8a5cb8a25c8bb4f687ca44b583a68d4aL57-R64):
Added "Run CodeLimit" to the list of required checks.