Schedule Code Checks to Run Daily (#113)
# Pull Request

## Description

This pull request includes a modification to the GitHub Actions workflow
file to streamline the naming convention and add a scheduled run for
code checks.

Changes to the GitHub Actions workflow:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L1-R9):
Renamed the job from "Run Code Checks" to "Code Checks" and added a cron
schedule to run the job daily at midnight.