Update development-ideas status checks (#137)
# Pull Request

## Description

This pull request includes an addition to the branch protection rules in
the `stack/development-ideas.tf` file. The change adds a new required
status check to ensure the Justfile format is validated.

*
[`stack/development-ideas.tf`](diffhunk://#diff-b235d4ce59faa805849e113c98ed0576b335ba180dbc2db6a06ce02b1beda850R55):
Added "Check Justfile Format" to the list of required status checks in
the `development-ideas_default_branch_protection` module.