Update repository-visualiser status checks (#143)
# Pull Request

## Description

This pull request includes updates to the
`stack/repository-visualiser.tf` file to enhance the repository
visualizer's default branch protection settings. The most important
changes are the addition of new checks to the list of required status
checks.

Enhancements to default branch protection:

*
[`stack/repository-visualiser.tf`](diffhunk://#diff-ce6778ad24df568bf231f7fefa425873c3c9de592ed6fc6799151f28999a17b7R56-R58):
Added "Check Justfile Format" and "Repository Visualiser" to the list of
required status checks.