Update Dependabot Configuration (#146)
# Pull Request

## Description

This pull request includes changes to the Dependabot configuration file
to adjust the update schedule for dependencies. The most important
change is the modification of the update interval from weekly to daily.

Changes to Dependabot schedule:

*
[`.github/dependabot.yml`](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L8-R8):
Changed the update interval from "weekly" to "daily" for the root
directory and specified subdirectories.
[[1]](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L8-R8)
[[2]](diffhunk://#diff-dd4fbda47e51f1e35defb9275a9cd9c212ecde0b870cba89ddaaae65c5f3cd28L26-R25)