Add Lefthook Validate Job (#164)
# Pull Request

## Description

This pull request adds a new job to the GitHub Actions workflow for code
checks. The new job, named "Lefthook Validate," is designed to validate
the code using Lefthook.

New job addition:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4R96-R111):
Added a new job `lefthook-validate` to the GitHub Actions workflow. This
job checks out the repository, installs the latest version of `uv`, and
runs `uvx lefthook validate` to validate the code.