Update tech-radar status checks (#166)
# Pull Request

## Description

This pull request includes a minor addition to the `stack/tech-radar.tf`
file. The change adds a new validation step to the list of required
status checks for the `tech-radar_default_branch_protection` module.

*
[`stack/tech-radar.tf`](diffhunk://#diff-492e56ddb82f7c2f0feda77f5a318465a9a443db5775cad33ae6161905eecc50R64):
Added "Lefthook Validate" to the list of required status checks in the
`tech-radar_default_branch_protection` module.