Add SlocCount repository description (#211)
# Pull Request

## Description

This pull request includes a change to the `stack/SlocCount.tf` file,
where the description of the `github_repository` resource has been
updated to provide more context about the repository.

*
[`stack/SlocCount.tf`](diffhunk://#diff-d739447262e9ffe0d94173aa0fbecf900d26abd606d622fa121f67b5cbc87936L5-R5):
Updated the `description` field for the `github_repository` resource to
"A repository to count the lines of code in a project" to provide a
clearer explanation of the repository's purpose.