Add GitHub Action Step Name (#214)
# Pull Request

## Description

This pull request makes a minor improvement to the GitHub Actions
workflow by adding a descriptive name to the label synchronization step
in the `.github/workflows/sync-labels.yml` file.

*
[`.github/workflows/sync-labels.yml`](diffhunk://#diff-a877ed9f27d115d95934fd904f2475dcec6ce4125da686dd5b3c75a696fff1c6L25-R26):
Updated the `micnncim/action-label-syncer` step to include a `name`
field, specifying it as "Sync labels" for better readability in the
workflow logs.