Pinact Check Job (#219)
# Pull Request

## Description

This pull request introduces configuration and workflow updates to
integrate and verify the use of Pinact, a tool for managing pinned
dependencies, as well as minor updates to existing configurations. Below
is a breakdown of the most important changes:

### Pinact Integration:

* Added a new Pinact configuration file
`.github/other-configurations/pinact.yml` to define the version and
specify actions to ignore (`actions/.*` and `github/codeql-action/.*`).
* Updated `Justfile` to include new commands for running and verifying
Pinact (`pinact-run` and `pinact-check`).
* Modified `lefthook.yml` to add a new pre-commit hook (`Pinact Checks`)
that runs the `pinact-check` command.

### Workflow Updates:

* Added a new `pinact-verify` job to `.github/workflows/code-checks.yml`
to verify Pinact configurations. This job includes steps for checking
out the repository and installing the latest version of `uv`.
* Updated the `uv` installation step in the existing workflow to use a
specific commit hash instead of a version tag for better
reproducibility.

### Configuration Updates:

* Increased the `lefthook` minimum version in `lefthook.yml` from
`1.11.3` to `1.11.11` to ensure compatibility with the new hooks.