Bump GitHub Action Versions (#221)
# Pull Request

## Description

This pull request updates dependencies in the GitHub Actions workflow
and introduces a new command in the `Justfile`. The most significant
changes include upgrading the versions of external actions used in the
workflow and adding a `pinact-update` command for easier updates.

### Workflow dependency updates:
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L50-R50):
Updated the `UmbrellaDocs/action-linkspector` action to version `v1.3.4`
for checking Markdown links.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L85-R85):
Updated the `astral-sh/setup-uv` action to version `v5.4.2` for
installing the latest version of `uv`.
[[1]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L85-R85)
[[2]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L108-R108)

### New command addition:
*
[`Justfile`](diffhunk://#diff-2f90408c2b0302b1cdb7f5d634114750837f940fa82d13057d9c18d0170a7e5cR85-R88):
Added a `pinact-update` command to run `pinact` with the `--update`
option, streamlining the process of updating configurations.