Update RepoOrchestrator status checks (#231)
# Pull Request

## Description

This pull request removes the deprecated "Pinact Verify" workflow and
updates the associated configurations to reflect this change.

### Workflow updates:

*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L132-L145):
Removed the "Pinact Verify" job, which included steps for repository
checkout and installing the latest version of `uv`.

### Configuration updates:

*
[`stack/RepoOrchestrator.tf`](diffhunk://#diff-91f35e38e4fb154b436ffff883598acbcd0280c0516ddbbfa24d768525b9e306L51):
Removed "Pinact Verify" from the list of required status checks in the
`RepoOrchestrator_default_branch_protection` module.