Update RepoOrchestrator status checks (#283)
# Pull Request

## Description

This pull request updates the required status checks for the default
branch protection in the `RepoOrchestrator` module to align with the new
naming convention and include an additional check.

Updates to default branch protection:

*
[`stack/RepoOrchestrator.tf`](diffhunk://#diff-91f35e38e4fb154b436ffff883598acbcd0280c0516ddbbfa24d768525b9e306L44-L50):
Updated the names of several status checks to use the "Common Code
Checks /" prefix and added a new check, "Common Code Checks / Lefthook
Validate," to the list of required status checks. The standalone
"Lefthook Validate" check was removed to avoid duplication.