Update SlocCount status checks (#298)
# Pull Request

## Description

This pull request updates the default branch protection rules in the
`SlocCount` module to align with a new naming convention for status
checks and includes the addition of a previously omitted check.

### Updates to branch protection rules:

* Renamed several status checks to follow the new "Common Code Checks"
naming convention, such as `"Check GitHub Actions with zizmor"` to
`"Common Code Checks / Check GitHub Actions with zizmor"`.
* Added the `"Common Code Checks / Lefthook Validate"` status check,
which was previously missing from the branch protection rules.