Add projects repository description (#336)
# Pull Request

## Description

This pull request updates the descriptions of two GitHub repositories in
Terraform configuration files to make them more descriptive and
user-friendly.

Changes to repository descriptions:

*
[`stack/github-stats.tf`](diffhunk://#diff-4c83331edafb5463d9be52eca7c9e213c0868cddd414c2b7d97d4e1559a70ab1L5-R5):
Updated the description of the `github-stats` repository to "Website to
display my Repository Statistics" for improved clarity.
*
[`stack/projects.tf`](diffhunk://#diff-b0a37de59757357b8080f1ff625c99785318d1e20d1679ea7caeeafc0a079774L5-R5):
Added a description to the `projects` repository, specifying it as "A
repository to list my projects and their status."