Update SlocCount status checks (#340)
# Pull Request

## Description

This pull request updates the configuration for the `SlocCount` GitHub
repository to include a homepage URL and enable GitHub Pages. These
changes enhance the repository's accessibility and hosting capabilities.

Enhancements to repository configuration:

*
[`stack/SlocCount.tf`](diffhunk://#diff-d739447262e9ffe0d94173aa0fbecf900d26abd606d622fa121f67b5cbc87936R7):
Added a `homepage_url` pointing to the repository's GitHub Pages site to
improve visibility and accessibility.
*
[`stack/SlocCount.tf`](diffhunk://#diff-d739447262e9ffe0d94173aa0fbecf900d26abd606d622fa121f67b5cbc87936R37-R45):
Configured GitHub Pages with a `build_type` set to "workflow" and
specified the source branch and path, enabling the repository to serve a
static site.