Update RepoOrchestrator status checks (#357)
# Pull Request

## Description

This pull request updates the branch protection rules in the
`RepoOrchestrator` Terraform module. The most notable change is the
modification of a required status check to reflect a more detailed
naming convention.

### Updates to branch protection rules:
*
[`stack/RepoOrchestrator.tf`](diffhunk://#diff-91f35e38e4fb154b436ffff883598acbcd0280c0516ddbbfa24d768525b9e306L44-R44):
Updated the required status check `"CodeQL Analysis"` to `"CodeQL
Analysis (actions) / Analyse code"` for better clarity and alignment
with naming conventions.