Migrate to Reusable CodeQL Workflow (#344)
# Pull Request

## Description

This pull request updates the CodeQL analysis workflow in
`.github/workflows/code-checks.yml` to use a reusable workflow,
simplifying the configuration and improving maintainability.

### Workflow simplification:

* Renamed the job from `run-codeql-analysis` to `codeql-checks` for
clarity and consistency.
* Replaced inline steps for CodeQL initialization and analysis with a
reusable workflow from `JackPlowman/reusable-workflows`. This reduces
redundancy and centralizes the workflow logic.
* Introduced a matrix strategy for specifying the `language` parameter,
allowing for easier expansion to support additional languages in the
future.
* Adjusted permissions to explicitly set `contents: read` and
`security-events: write`, aligning with the principle of least
privilege.