Update tech-radar status checks (#358)
# Pull Request

## Description

This pull request updates the branch protection rules for the
`tech-radar` repository. The most important change is a modification to
the required status checks.

Branch protection updates:

*
[`stack/tech-radar.tf`](diffhunk://#diff-492e56ddb82f7c2f0feda77f5a318465a9a443db5775cad33ae6161905eecc50L58-R58):
Updated the `required_status_checks` list to replace `"CodeQL Analysis"`
with `"CodeQL Analysis (actions) / Analyse code"` to reflect the new
naming convention for this check.