Update project-links status checks (#364)
# Pull Request

## Description

This pull request updates the branch protection rules for the
`project-links` repository by modifying the required status checks. The
changes ensure that the status check names include more descriptive
labels for better clarity.

### Updates to branch protection rules:

*
[`stack/project-links.tf`](diffhunk://#diff-8d4a200df5e9c4325ae12ffba43bb709aed7403c35680e7b6b89e058fa1f37aeL61-R63):
Updated the names of the `CodeQL Analysis` status checks to include the
`/ Analyse code` suffix for the `actions`, `python`, and `typescript`
checks.