Update Reusable Workflows (#426)
# Pull Request

## Description

This pull request updates the reusable workflows referenced in several
GitHub Actions workflow files to a newer version. The changes ensure
that the workflows use the latest functionality and fixes provided in
version `v2025.07.01.03`.

Updated reusable workflow references:

*
[`.github/workflows/clean-caches.yml`](diffhunk://#diff-d0394e4336a74cdfc1d4cff05d056b893ac7ff922eacf4448e104a754f386b8dL15-R15):
Updated the reference to `common-clean-caches.yml` to version
`v2025.07.01.03`.
*
[`.github/workflows/code-checks.yml`](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L48-R48):
Updated the references to both `common-code-checks.yml` and
`codeql-analysis.yml` to version `v2025.07.01.03`.
[[1]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L48-R48)
[[2]](diffhunk://#diff-ddf88e15b08104435ae66be9982938335f6c290a85de4cb9a09868e0e01dd4d4L60-R60)
*
[`.github/workflows/pull-request-tasks.yml`](diffhunk://#diff-ba6496a5b7a58ac3681ed047691dc32281cc7d548fff1d41201babbd65ad45cfL14-R14):
Updated the reference to `common-pull-request-tasks.yml` to version
`v2025.07.01.03`.
*
[`.github/workflows/sync-labels.yml`](diffhunk://#diff-a877ed9f27d115d95934fd904f2475dcec6ce4125da686dd5b3c75a696fff1c6L19-R19):
Updated the reference to `common-sync-labels.yml` to version
`v2025.07.01.03`.